Add basic support for UB players
Modern Panasonic UB players support a similar interface to the old BD
Panasonic range. To do most operations seems to involve some sort of
authentication, but we can enable a subclass of status operations which
at least allows some retrieval of state.

Default to auto-detecting the player type, setting the type
appropriately when we see a success or failure from a call. Based on
initial work by Dan Smith.

Fixes #1